pyproject.toml
[project]
name = "issue-agent"
version = "0.1.0"
description = "AI agent-based personalized issue notification system"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "praw>=7.7.0",
    "browser-use>=0.1.0",
    "undetected-chromedriver>=3.5.0",
    "selenium>=4.15.0",
    "playwright>=1.40.0",
    "python-dateutil>=2.8.2",
    "urllib3>=2.0.0",
    "python-dotenv>=1.0.0",
    "ratelimit>=2.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "models", "config", "utils"]
